Extract templates registration to separate module
